+++ src/main/kotlin/com/coder/toolbox/store/StoreKeys.kt
@@ -2,6 +2,8 @@ package com.coder.toolbox.store
 
 internal const val CODER_SSH_CONFIG_OPTIONS = "CODER_SSH_CONFIG_OPTIONS"
 
+internal const val CODER_HEADER_COMMAND = "CODER_HEADER_COMMAND"
+
 internal const val LAST_USED_URL = "lastDeploymentURL"
 
 internal const val DEFAULT_URL = "defaultURL"