+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.9.4 - 2026-08-26
+
 ### Added
 
 - support for configuring the SSH config path, letting Toolbox manage a separate SSH config file instead of always